requirements.txt
pyTelegramBotAPI==3.8.1
Flask==2.0.1
Pillow==8.3.1
telegram==0.0.1
emoji==1.4.1
python-telegram-bot==13.7
numpy==1.21.1